pyproject.toml
[project]
name = "shallow-faker"
version = "0.1.0"
description = "Automated pipeline for creating Piper-compatible voice clones"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "faster-whisper>=1.0.0",
    "piper-phonemize>=1.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "ffmpeg-python>=0.2.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
shallow-fake = "shallow_fake.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["shallow_fake"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

